+++ packages/twenty-server/src/modules/calendar-messaging-participant/utils/get-company-name-from-domain-name.util.ts
