+++ packages/twenty-server/src/modules/calendar-messaging-participant/utils/get-domain-name-from-handle.util.ts
