+++ packages/twenty-server/src/modules/calendar-messaging-participant/utils/get-first-name-and-last-name-from-handle-and-display-name.util.ts
