+++ packages/twenty-server/src/modules/calendar-messaging-participant/utils/is-person-email.util.ts
