+++ packages/twenty-server/src/modules/calendar/crons/commands/calendar-cron-commands.module.ts
@@ -1,8 +1,5 @@
 import { Module } from '@nestjs/common';
 
-import { GoogleCalendarSyncCommand } from 'src/modules/calendar/commands/google-calendar-sync.command';
-import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repository/object-metadata-repository.module';
-import { ConnectedAccountObjectMetadata } from 'src/modules/connected-account/standard-objects/connected-account.object-metadata';
 import { GoogleCalendarSyncCronCommand } from 'src/modules/calendar/crons/commands/google-calendar-sync.cron.command';
 
 @Module({
